+++ src/main/java/de/jeff_media/AngelChest/ConfigUtils.java
@@ -60,6 +60,8 @@ public class ConfigUtils {
 		main.getConfig().addDefault("material", "CHEST");
 		main.getConfig().addDefault("player-head","{PLAYER}");
 		main.getConfig().addDefault("preserve-xp", true);
+		main.getConfig().addDefault("remove-curse-of-vanishing",true);
+		main.getConfig().addDefault("remove-curse-of-binding",true);
 		main.getConfig().addDefault("full-xp", false); // Currently not in config because there is no way to get players XP
 		main.disabledWorlds = (ArrayList<String>) main.getConfig().getStringList("disabled-worlds");